requirements.txt
# These requirements are for the Python that is run in the
# build container.
#
# Requirements for tests, code style and other build stuff reside in
# requirements-build.txt

# In-house dependencies
leaf-common>=1.2.31

# Other dependencies

# These are needed for grpc integration
grpcio>=1.62.0
grpcio-health-checking>=1.62.0
grpcio-reflection>=1.62.0
grpcio-tools>=1.62.0

# Open-telemetry logging
opentelemetry-api>=1.39.0
opentelemetry-sdk>=1.39.0
opentelemetry-exporter-otlp>=1.39.0
